monitor/requirements.txt
asttokens==2.1.0
backcall==0.2.0
decorator==5.1.1
executing==1.2.0
ipython==8.6.0
jedi==0.18.1
matplotlib-inline==0.1.6
mysql-connector-python==8.0.29
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.32
protobuf==4.21.9
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.13.0
python-dotenv==0.21.0
six==1.16.0
stack-data==0.6.0
traitlets==5.5.0
wcwidth==0.2.5
pudb==2022.1.3
